requirements.txt
python-telegram-bot
python-dotenv
SQLAlchemy
Pillow